x86: merge IAMCU and SYS V core arch code
Having two parallel implementations is a maintenance issue, especially
when some strategically placed #ifdefs will suffice.

We prefer the ASM versions for SYS V, as we need complete control of
the emitted assembly for interrupt handling and context switching.
The SYS V code is far more mature. IAMCU C code has known issues with
-fomit-frame-pointer.

The only difference between the two calling conventions is that the
first three function arguments are provided in eax, edx, ecx instead
of on the stack.

Issue: ZEP-49

Former-commit-id: 8c524a291e7ee20d031b71653b312c993fcadf74

+++ arch/x86/core/intconnect.c
@@ -5,7 +5,7 @@ DECL|_IntVecSet|function|void _IntVecSet(unsigned int vector, void (*routine)(vo
 DECL|_SpuriousIntHandler|variable|_SpuriousIntHandler
 DECL|_SpuriousIntNoErrCodeHandler|variable|_SpuriousIntNoErrCodeHandler
 DECL|_arch_irq_connect_dynamic|function|int _arch_irq_connect_dynamic(unsigned int irq, unsigned int priority,void (*routine)(void *parameter), void *parameter, uint32_t flags)
-DECL|_common_dynamic_irq_handler|function|void _common_dynamic_irq_handler(uint32_t stub_idx)
+DECL|_common_dynamic_irq_handler|function|void _common_dynamic_irq_handler(uint8_t stub_idx)
 DECL|_dummy_exception_vector_stub|variable|_dummy_exception_vector_stub
 DECL|_dummy_spurious_interrupt|variable|_dummy_spurious_interrupt
 DECL|_irq_handler_set|function|void _irq_handler_set(unsigned int vector, void (*routine)(void *parameter), void *parameter)
